requirements.txt
crewai
crewai-tools
textgrad
langchain
langchain-openai
python-dotenv